Fix a bug to support Float32

+++ tinyGP/runTinyGP.jl
@@ -49,7 +49,7 @@ function main(argv)
     objective = parsed["objective"]
     testdataset = parsed["test"] == "" ? trainingfilename : parsed["test"]
 
-    X_test, y_test = TinyGP.load_dataset(Float64, testdataset, targetname)
+    X_test, y_test = TinyGP.load_dataset(Float32, testdataset, targetname)
 
     loss_func = paramopt_loss_func = TinyGP.mean_squared_error # default
     
@@ -67,7 +67,7 @@ function main(argv)
     else
         error("unknown objective function value (allowed values are mse, r2, dl)")
     end
-    gp = TinyGP.Algorithm{Float64}(trainingfilename, targetname,
+    gp = TinyGP.Algorithm{Float32}(trainingfilename, targetname,
         generations=generations, popsize=popsize, maxlen=maxlen, tournamentsize=tsize, 
         loss_func=loss_func, paramopt_loss_func=paramopt_loss_func)
